removed ProducerCore (base)
