+++ src/Confluent.Kafka/Producer.cs
@@ -16,18 +16,29 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Confluent.Kafka.Serdes;
+using Confluent.Kafka.Impl;
+using Confluent.Kafka.Internal;
+
 
 namespace Confluent.Kafka
 {
     /// <summary>
     ///     A high level producer with serialization capability.
     /// </summary>
-    public class Producer<TKey, TValue> : IProducer<TKey, TValue>
+    public class Producer<TKey, TValue> : IProducer<TKey, TValue>, IClient
     {
-        private ProducerCore core;
+        internal class Config
+        {
+            internal IEnumerable<KeyValuePair<string, string>> config;
+            internal Action<Error> errorHandler;
+            internal Action<LogMessage> logHandler;
+            internal Action<string> statisticsHandler;
+        }
 
         private ISerializer<TKey> keySerializer;
         private ISerializer<TValue> valueSerializer;
@@ -45,15 +56,410 @@ namespace Confluent.Kafka
             { typeof(byte[]), Serializers.ByteArray }
         };
 
+        private int cancellationDelayMaxMs;
+        private bool disposeHasBeenCalled = false;
+        private object disposeHasBeenCalledLockObj = new object();
+
+        private bool manualPoll = false;
+        internal bool enableDeliveryReports = true;
+        internal bool enableDeliveryReportKey = true;
+        internal bool enableDeliveryReportValue = true;
+        internal bool enableDeliveryReportTimestamp = true;
+        internal bool enableDeliveryReportHeaders = true;
+        internal bool enableDeliveryReportPersistedStatus = true;
+
+        private SafeKafkaHandle ownedKafkaHandle;
+        private Handle borrowedHandle;
+
+        internal SafeKafkaHandle KafkaHandle
+            => ownedKafkaHandle != null 
+                ? ownedKafkaHandle
+                : borrowedHandle.LibrdkafkaHandle;
+
+        private Task callbackTask;
+        private CancellationTokenSource callbackCts;
+
+        private Task StartPollTask(CancellationToken ct)
+            => Task.Factory.StartNew(() =>
+                {
+                    try
+                    {
+                        while (true)
+                        {
+                            ct.ThrowIfCancellationRequested();
+                            ownedKafkaHandle.Poll((IntPtr)cancellationDelayMaxMs);
+                        }
+                    }
+                    catch (OperationCanceledException) {}
+                }, ct, TaskCreationOptions.LongRunning, TaskScheduler.Default);
+
+
+        private Action<Error> errorHandler;
+        private Librdkafka.ErrorDelegate errorCallbackDelegate;
+        private void ErrorCallback(IntPtr rk, ErrorCode err, string reason, IntPtr opaque)
+        {
+            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
+            if (ownedKafkaHandle.IsClosed) { return; }
+            errorHandler?.Invoke(KafkaHandle.CreatePossiblyFatalError(err, reason));
+        }
+
+
+        private Action<string> statisticsHandler;
+        private Librdkafka.StatsDelegate statisticsCallbackDelegate;
+        private int StatisticsCallback(IntPtr rk, IntPtr json, UIntPtr json_len, IntPtr opaque)
+        {
+            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
+            if (ownedKafkaHandle.IsClosed) { return 0; }
+            statisticsHandler?.Invoke(Util.Marshal.PtrToStringUTF8(json));
+            return 0; // instruct librdkafka to immediately free the json ptr.
+        }
+
+
+        private Action<LogMessage> logHandler;
+        private object loggerLockObj = new object();
+        private Librdkafka.LogDelegate logCallbackDelegate;
+        private void LogCallback(IntPtr rk, SyslogLevel level, string fac, string buf)
+        {
+            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
+            // Note: kafkaHandle can be null if the callback is during construction (in that case, we want the delegate to run).
+            if (ownedKafkaHandle != null && ownedKafkaHandle.IsClosed) { return; }
+            logHandler?.Invoke(new LogMessage(Util.Marshal.PtrToStringUTF8(Librdkafka.name(rk)), level, fac, buf));
+        }
+
+        private Librdkafka.DeliveryReportDelegate DeliveryReportCallback;
+
+        private void DeliveryReportCallbackImpl(IntPtr rk, IntPtr rkmessage, IntPtr opaque)
+        {
+            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
+            if (ownedKafkaHandle.IsClosed) { return; }
+
+            var msg = Util.Marshal.PtrToStructureUnsafe<rd_kafka_message>(rkmessage);
+
+            // the msg._private property has dual purpose. Here, it is an opaque pointer set
+            // by Topic.Produce to be an IDeliveryHandler. When Consuming, it's for internal
+            // use (hence the name).
+            if (msg._private == IntPtr.Zero)
+            {
+                // Note: this can occur if the ProduceAsync overload that accepts a DeliveryHandler
+                // was used and the delivery handler was set to null.
+                return;
+            }
+
+            var gch = GCHandle.FromIntPtr(msg._private);
+            var deliveryHandler = (IDeliveryHandler) gch.Target;
+            gch.Free();
+
+            Headers headers = null;
+            if (this.enableDeliveryReportHeaders) 
+            {
+                headers = new Headers();
+                Librdkafka.message_headers(rkmessage, out IntPtr hdrsPtr);
+                if (hdrsPtr != IntPtr.Zero)
+                {
+                    for (var i=0; ; ++i)
+                    {
+                        var err = Librdkafka.header_get_all(hdrsPtr, (IntPtr)i, out IntPtr namep, out IntPtr valuep, out IntPtr sizep);
+                        if (err != ErrorCode.NoError)
+                        {
+                            break;
+                        }
+                        var headerName = Util.Marshal.PtrToStringUTF8(namep);
+                        byte[] headerValue = null;
+                        if (valuep != IntPtr.Zero)
+                        {
+                            headerValue = new byte[(int)sizep];
+                            Marshal.Copy(valuep, headerValue, 0, (int)sizep);
+                        }
+                        headers.Add(headerName, headerValue);
+                    }
+                }
+            }
+
+            IntPtr timestampType = (IntPtr)TimestampType.NotAvailable;
+            long timestamp = 0;
+            if (enableDeliveryReportTimestamp)
+            {
+                timestamp = Librdkafka.message_timestamp(rkmessage, out timestampType);
+            }
+
+            PersistenceStatus messageStatus = PersistenceStatus.PossiblyPersisted;
+            if (enableDeliveryReportPersistedStatus)
+            {
+                messageStatus = Librdkafka.message_status(rkmessage);
+            }
+
+            deliveryHandler.HandleDeliveryReport(
+                new DeliveryReport<Null, Null>
+                {
+                    // Topic is not set here in order to avoid the marshalling cost.
+                    // Instead, the delivery handler is expected to cache the topic string.
+                    Partition = msg.partition, 
+                    Offset = msg.offset, 
+                    Error = KafkaHandle.CreatePossiblyFatalError(msg.err, null),
+                    Message = new Message<Null, Null> { Timestamp = new Timestamp(timestamp, (TimestampType)timestampType), Headers = headers }
+                }
+            );
+        }
+
+        private void ProduceImpl(
+            string topic,
+            byte[] val, int valOffset, int valLength,
+            byte[] key, int keyOffset, int keyLength,
+            Timestamp timestamp,
+            Partition partition, 
+            IEnumerable<Header> headers,
+            IDeliveryHandler deliveryHandler)
+        {
+            if (timestamp.Type != TimestampType.CreateTime)
+            {
+                if (timestamp != Timestamp.Default)
+                {
+                    throw new ArgumentException("Timestamp must be either Timestamp.Default, or Timestamp.CreateTime.");
+                }
+            }
+
+            if (this.enableDeliveryReports && deliveryHandler != null)
+            {
+                // Passes the TaskCompletionSource to the delivery report callback via the msg_opaque pointer
+
+                // Note: There is a level of indirection between the GCHandle and
+                // physical memory address. GCHandle.ToIntPtr doesn't get the
+                // physical address, it gets an id that refers to the object via
+                // a handle-table.
+                var gch = GCHandle.Alloc(deliveryHandler);
+                var ptr = GCHandle.ToIntPtr(gch);
+
+                var err = KafkaHandle.Produce(
+                    topic,
+                    val, valOffset, valLength,
+                    key, keyOffset, keyLength,
+                    partition.Value,
+                    timestamp.UnixTimestampMs,
+                    headers,
+                    ptr);
+
+                if (err != ErrorCode.NoError)
+                {
+                    gch.Free();
+                    throw new KafkaException(KafkaHandle.CreatePossiblyFatalError(err, null));
+                }
+            }
+            else
+            {
+                var err = KafkaHandle.Produce(
+                    topic,
+                    val, valOffset, valLength,
+                    key, keyOffset, keyLength,
+                    partition.Value,
+                    timestamp.UnixTimestampMs,
+                    headers,
+                    IntPtr.Zero);
+
+                if (err != ErrorCode.NoError)
+                {
+                    throw new KafkaException(KafkaHandle.CreatePossiblyFatalError(err, null));
+                }
+            }
+        }
+
+
         /// <summary>
-        ///     <see cref="IClient.Handle" />
+        ///     Poll for callback events. Typically, you should not 
+        ///     call this method. Only call on producer instances 
+        ///     where background polling has been disabled.
         /// </summary>
-        public Handle Handle => core.Handle;
+        /// <param name="timeout">
+        ///     The maximum period of time to block if no callback events
+        ///     are waiting. You should typically use a relatively short 
+        ///     timout period because this operation cannot be cancelled.
+        /// </param>
+        /// <returns>
+        ///     Returns the number of events served.
+        /// </returns>
+        public int Poll(TimeSpan timeout)
+        {
+            if (!manualPoll)
+            {
+                throw new InvalidOperationException("Poll method called, but manual polling is not enabled.");
+            }
+
+            return this.KafkaHandle.Poll((IntPtr)timeout.TotalMillisecondsAsInt());
+        }
+
+
+        /// <summary>
+        ///     Wait until all outstanding produce requests and delievery report
+        ///     callbacks are completed.
+        ///    
+        ///     [API-SUBJECT-TO-CHANGE] - the semantics and/or type of the return value
+        ///     is subject to change.
+        /// </summary>
+        /// <param name="timeout">
+        ///     The maximum length of time to block. You should typically use a
+        ///     relatively short timout period and loop until the return value
+        ///     becomes zero because this operation cannot be cancelled. 
+        /// </param>
+        /// <returns>
+        ///     The current librdkafka out queue length. This should be interpreted
+        ///     as a rough indication of the number of messages waiting to be sent
+        ///     to or acknowledged by the broker. If zero, there are no outstanding
+        ///     messages or callbacks. Specifically, the value is equal to the sum
+        ///     of the number of produced messages for which a delivery report has
+        ///     not yet been handled and a number which is less than or equal to the
+        ///     number of pending delivery report callback events (as determined by
+        ///     the number of outstanding protocol requests).
+        /// </returns>
+        /// <remarks>
+        ///     This method should typically be called prior to destroying a producer
+        ///     instance to make sure all queued and in-flight produce requests are
+        ///     completed before terminating. The wait time is bounded by the
+        ///     timeout parameter.
+        ///    
+        ///     A related configuration parameter is message.timeout.ms which determines
+        ///     the maximum length of time librdkafka attempts to deliver a message 
+        ///     before giving up and so also affects the maximum time a call to Flush 
+        ///     may block.
+        /// 
+        ///     Where this Producer instance shares a Handle with one or more other
+        ///     producer instances, the Flush method will wait on messages produced by
+        ///     the other producer instances as well.
+        /// </remarks>
+        public int Flush(TimeSpan timeout)
+            => KafkaHandle.Flush(timeout.TotalMillisecondsAsInt());
+
+
+        /// <summary>
+        ///     Wait until all outstanding produce requests and delievery report
+        ///     callbacks are completed.
+        /// </summary>
+        /// <remarks>
+        ///     This method should typically be called prior to destroying a producer
+        ///     instance to make sure all queued and in-flight produce requests are
+        ///     completed before terminating. 
+        ///    
+        ///     A related configuration parameter is message.timeout.ms which determines
+        ///     the maximum length of time librdkafka attempts to deliver a message 
+        ///     before giving up and so also affects the maximum time a call to Flush 
+        ///     may block.
+        /// 
+        ///     Where this Producer instance shares a Handle with one or more other
+        ///     producer instances, the Flush method will wait on messages produced by
+        ///     the other producer instances as well.
+        /// </remarks>
+        /// <exception cref="System.OperationCanceledException">
+        ///     Thrown if the operation is cancelled.
+        /// </exception>
+        public void Flush(CancellationToken cancellationToken)
+        {
+            while (true)
+            {
+                int result = KafkaHandle.Flush(100);
+                if (result == 0)
+                {
+                    return;
+                }
+                if (cancellationToken.IsCancellationRequested)
+                {
+                    // TODO: include flush number in exception.
+                    throw new OperationCanceledException();
+                }
+            }
+        }
+
+        
+        /// <summary>
+        ///     Releases all resources used by this <see cref="Producer{TKey,TValue}" />.
+        /// </summary>
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+
+
+        /// <summary>
+        ///     Releases the unmanaged resources used by the
+        ///     <see cref="Producer{TKey,TValue}" />
+        ///     and optionally disposes the managed resources.
+        /// </summary>
+        /// <param name="disposing">
+        ///     true to release both managed and unmanaged resources;
+        ///     false to release only unmanaged resources.
+        /// </param>
+        protected virtual void Dispose(bool disposing)
+        {
+            // Calling Dispose a second or subsequent time should be a no-op.
+            lock (disposeHasBeenCalledLockObj)
+            { 
+                if (disposeHasBeenCalled) { return; }
+                disposeHasBeenCalled = true;
+            }
+
+            // do nothing if we borrowed a handle.
+            if (ownedKafkaHandle == null) { return; }
+
+            if (disposing)
+            {
+                if (!this.manualPoll)
+                {
+                    callbackCts.Cancel();
+                    try
+                    {
+                        // Note: It's necessary to wait on callbackTask before disposing kafkaHandle
+                        // since the poll loop makes use of this.
+                        callbackTask.Wait();
+                    }
+                    catch (AggregateException e)
+                    {
+                        if (e.InnerException.GetType() != typeof(TaskCanceledException))
+                        {
+                            throw e.InnerException;
+                        }
+                    }
+                    finally
+                    {
+                        callbackCts.Dispose();
+                    }
+                }
+
+                // calls to rd_kafka_destroy may result in callbacks
+                // as a side-effect. however the callbacks this class
+                // registers with librdkafka ensure that any registered
+                // events are not called if kafkaHandle has been closed.
+                // this avoids deadlocks in common scenarios.
+                ownedKafkaHandle.Dispose();
+            }
+        }
 
         /// <summary>
         ///     <see cref="IClient.Name" />
         /// </summary>
-        public string Name => core.Name;
+        public string Name
+            => KafkaHandle.Name;
+
+
+        /// <summary>
+        ///     <see cref="IClient.AddBrokers(string)" />
+        /// </summary>
+        public int AddBrokers(string brokers)
+            => KafkaHandle.AddBrokers(brokers);
+
+
+        /// <summary>
+        ///     <see cref="IClient.Handle" />
+        /// </summary>
+        public Handle Handle 
+        {
+            get
+            {
+                if (this.ownedKafkaHandle != null)
+                {
+                    return new Handle { Owner = this, LibrdkafkaHandle = ownedKafkaHandle };
+                }
+                
+                return borrowedHandle;
+            }
+        }
 
         private void InitializeSerializers(
             ISerializer<TKey> keySerializer,
@@ -110,7 +516,13 @@ namespace Confluent.Kafka
 
         internal Producer(DependentProducerBuilder<TKey, TValue> builder)
         {
-            core = new ProducerCore(builder.Handle);
+            this.borrowedHandle = builder.Handle;
+
+            if (!borrowedHandle.Owner.GetType().Name.Contains("Producer")) // much simpler than checking actual types.
+            {
+                throw new Exception("A Producer instance may only be constructed using the handle of another Producer instance.");
+            }
+
             InitializeSerializers(
                 builder.KeySerializer, builder.ValueSerializer,
                 builder.AsyncKeySerializer, builder.AsyncValueSerializer);
@@ -118,11 +530,120 @@ namespace Confluent.Kafka
 
         internal Producer(ProducerBuilder<TKey, TValue> builder)
         {
-            core = new ProducerCore(builder.ConstructBaseConfig(this));
+            var baseConfig = builder.ConstructBaseConfig(this);
+
+            // TODO: Make Tasks auto complete when EnableDeliveryReportsPropertyName is set to false.
+            // TODO: Hijack the "delivery.report.only.error" configuration parameter and add functionality to enforce that Tasks 
+            //       that never complete are never created when this is set to true.
+
+            this.statisticsHandler = baseConfig.statisticsHandler;
+            this.logHandler = baseConfig.logHandler;
+            this.errorHandler = baseConfig.errorHandler;
+
+            var config = Confluent.Kafka.Config.ExtractCancellationDelayMaxMs(baseConfig.config, out this.cancellationDelayMaxMs);
+
+            this.DeliveryReportCallback = DeliveryReportCallbackImpl;
+
+            Librdkafka.Initialize(null);
+
+            var modifiedConfig = config
+                .Where(prop => 
+                    prop.Key != ConfigPropertyNames.Producer.EnableBackgroundPoll &&
+                    prop.Key != ConfigPropertyNames.Producer.EnableDeliveryReports &&
+                    prop.Key != ConfigPropertyNames.Producer.DeliveryReportFields);
+
+            if (modifiedConfig.Where(obj => obj.Key == "delivery.report.only.error").Count() > 0)
+            {
+                // A managed object is kept alive over the duration of the produce request. If there is no
+                // delivery report generated, there will be a memory leak. We could possibly support this 
+                // property by keeping track of delivery reports in managed code, but this seems like 
+                // more trouble than it's worth.
+                throw new ArgumentException("The 'delivery.report.only.error' property is not supported by this client");
+            }
+
+            var enableBackgroundPollObj = config.FirstOrDefault(prop => prop.Key == ConfigPropertyNames.Producer.EnableBackgroundPoll).Value;
+            if (enableBackgroundPollObj != null)
+            {
+                this.manualPoll = !bool.Parse(enableBackgroundPollObj.ToString());
+            }
+
+            var enableDeliveryReportsObj = config.FirstOrDefault(prop => prop.Key == ConfigPropertyNames.Producer.EnableDeliveryReports).Value;
+            if (enableDeliveryReportsObj != null)
+            {
+                this.enableDeliveryReports = bool.Parse(enableDeliveryReportsObj.ToString());
+            }
+
+            var deliveryReportEnabledFieldsObj = config.FirstOrDefault(prop => prop.Key == ConfigPropertyNames.Producer.DeliveryReportFields).Value;
+            if (deliveryReportEnabledFieldsObj != null)
+            {
+                var fields = deliveryReportEnabledFieldsObj.ToString().Replace(" ", "");
+                if (fields != "all")
+                {
+                    this.enableDeliveryReportKey = false;
+                    this.enableDeliveryReportValue = false;
+                    this.enableDeliveryReportHeaders = false;
+                    this.enableDeliveryReportTimestamp = false;
+                    this.enableDeliveryReportPersistedStatus = false;
+                    if (fields != "none")
+                    {
+                        var parts = fields.Split(',');
+                        foreach (var part in parts)
+                        {
+                            switch (part)
+                            {
+                                case "key": this.enableDeliveryReportKey = true; break;
+                                case "value": this.enableDeliveryReportValue = true; break;
+                                case "timestamp": this.enableDeliveryReportTimestamp = true; break;
+                                case "headers": this.enableDeliveryReportHeaders = true; break;
+                                case "status": this.enableDeliveryReportPersistedStatus = true; break;
+                                default: throw new ArgumentException(
+                                    $"Unknown delivery report field name '{part}' in config value '{ConfigPropertyNames.Producer.DeliveryReportFields}'.");
+                            }
+                        }
+                    }
+                }
+            }
+
+            var configHandle = SafeConfigHandle.Create();
+
+            modifiedConfig.ToList().ForEach((kvp) => {
+                if (kvp.Value == null) throw new ArgumentException($"'{kvp.Key}' configuration parameter must not be null.");
+                configHandle.Set(kvp.Key, kvp.Value.ToString());
+            });
+
+
+            IntPtr configPtr = configHandle.DangerousGetHandle();
+
+            if (enableDeliveryReports)
+            {
+                Librdkafka.conf_set_dr_msg_cb(configPtr, DeliveryReportCallback);
+            }
+
+            // Explicitly keep references to delegates so they are not reclaimed by the GC.
+            errorCallbackDelegate = ErrorCallback;
+            logCallbackDelegate = LogCallback;
+            statisticsCallbackDelegate = StatisticsCallback;
+
+            // TODO: provide some mechanism whereby calls to the error and log callbacks are cached until
+            //       such time as event handlers have had a chance to be registered.
+            Librdkafka.conf_set_error_cb(configPtr, errorCallbackDelegate);
+            Librdkafka.conf_set_log_cb(configPtr, logCallbackDelegate);
+            Librdkafka.conf_set_stats_cb(configPtr, statisticsCallbackDelegate);
+
+            this.ownedKafkaHandle = SafeKafkaHandle.Create(RdKafkaType.Producer, configPtr, this);
+            configHandle.SetHandleAsInvalid(); // config object is no longer useable.
+
+            if (!manualPoll)
+            {
+                callbackCts = new CancellationTokenSource();
+                callbackTask = StartPollTask(callbackCts.Token);
+            }
+
             InitializeSerializers(
                 builder.KeySerializer, builder.ValueSerializer,
                 builder.AsyncKeySerializer, builder.AsyncValueSerializer);
         }
+
 
         /// <summary>
         ///     Asynchronously send a single message to a Kafka topic/partition.
@@ -159,15 +680,15 @@ namespace Confluent.Kafka
                     .GetAwaiter()
                     .GetResult();
 
-            if (core.enableDeliveryReports)
+            if (enableDeliveryReports)
             {
                 var handler = new TypedTaskDeliveryHandlerShim<TKey, TValue>(topicPartition.Topic,
-                    core.enableDeliveryReportKey ? message.Key : default(TKey),
-                    core.enableDeliveryReportValue ? message.Value : default(TValue));
+                    enableDeliveryReportKey ? message.Key : default(TKey),
+                    enableDeliveryReportValue ? message.Value : default(TValue));
 
                 cancellationToken.Register(() => handler.TrySetException(new TaskCanceledException()));
 
-                core.Produce(
+                ProduceImpl(
                     topicPartition.Topic,
                     valBytes, 0, valBytes == null ? 0 : valBytes.Length,
                     keyBytes, 0, keyBytes == null ? 0 : keyBytes.Length,
@@ -178,7 +699,7 @@ namespace Confluent.Kafka
             }
             else
             {
-                core.Produce(
+                ProduceImpl(
                     topicPartition.Topic, 
                     valBytes, 0, valBytes == null ? 0 : valBytes.Length, 
                     keyBytes, 0, keyBytes == null ? 0 : keyBytes.Length, 
@@ -265,7 +786,7 @@ namespace Confluent.Kafka
             Message<TKey, TValue> message,
             Action<DeliveryReport<TKey, TValue>> deliveryHandler = null)
         {
-            if (deliveryHandler != null && !core.enableDeliveryReports)
+            if (deliveryHandler != null && !enableDeliveryReports)
             {
                 throw new ArgumentException("A delivery handler was specified, but delivery reports are disabled.");
             }
@@ -284,7 +805,7 @@ namespace Confluent.Kafka
                     .GetAwaiter()
                     .GetResult();
 
-            core.Produce(
+            ProduceImpl(
                 topicPartition.Topic,
                 valBytes, 0, valBytes == null ? 0 : valBytes.Length, 
                 keyBytes, 0, keyBytes == null ? 0 : keyBytes.Length, 
@@ -292,103 +813,11 @@ namespace Confluent.Kafka
                 message.Headers, 
                 new TypedDeliveryHandlerShim_Action<TKey, TValue>(
                     topicPartition.Topic,
-                    core.enableDeliveryReportKey ? message.Key : default(TKey),
-                    core.enableDeliveryReportValue ? message.Value : default(TValue),
+                    enableDeliveryReportKey ? message.Key : default(TKey),
+                    enableDeliveryReportValue ? message.Value : default(TValue),
                     deliveryHandler)
             );
         }
-
-        /// <summary>
-        ///     Poll for callback events. Typically, you should not 
-        ///     call this method. Only call on producer instances 
-        ///     where background polling has been disabled.
-        /// </summary>
-        /// <param name="timeout">
-        ///     The maximum period of time to block if no callback events
-        ///     are waiting. You should typically use a relatively short 
-        ///     timout period because this operation cannot be cancelled.
-        /// </param>
-        /// <returns>
-        ///     Returns the number of events served.
-        /// </returns>
-        public int Poll(TimeSpan timeout)
-            => core.Poll(timeout);
-
-        /// <summary>
-        ///     Wait until all outstanding produce requests and delievery report
-        ///     callbacks are completed.
-        ///    
-        ///     [API-SUBJECT-TO-CHANGE] - the semantics and/or type of the return value
-        ///     is subject to change.
-        /// </summary>
-        /// <param name="timeout">
-        ///     The maximum length of time to block. You should typically use a
-        ///     relatively short timout period and loop until the return value
-        ///     becomes zero because this operation cannot be cancelled. 
-        /// </param>
-        /// <returns>
-        ///     The current librdkafka out queue length. This should be interpreted
-        ///     as a rough indication of the number of messages waiting to be sent
-        ///     to or acknowledged by the broker. If zero, there are no outstanding
-        ///     messages or callbacks. Specifically, the value is equal to the sum
-        ///     of the number of produced messages for which a delivery report has
-        ///     not yet been handled and a number which is less than or equal to the
-        ///     number of pending delivery report callback events (as determined by
-        ///     the number of outstanding protocol requests).
-        /// </returns>
-        /// <remarks>
-        ///     This method should typically be called prior to destroying a producer
-        ///     instance to make sure all queued and in-flight produce requests are
-        ///     completed before terminating. The wait time is bounded by the
-        ///     timeout parameter.
-        ///    
-        ///     A related configuration parameter is message.timeout.ms which determines
-        ///     the maximum length of time librdkafka attempts to deliver a message 
-        ///     before giving up and so also affects the maximum time a call to Flush 
-        ///     may block.
-        /// 
-        ///     Where this Producer instance shares a Handle with one or more other
-        ///     producer instances, the Flush method will wait on messages produced by
-        ///     the other producer instances as well.
-        /// </remarks>
-        public int Flush(TimeSpan timeout)
-            => core.Flush(timeout);
-
-        /// <summary>
-        ///     Wait until all outstanding produce requests and delievery report
-        ///     callbacks are completed.
-        /// </summary>
-        /// <remarks>
-        ///     This method should typically be called prior to destroying a producer
-        ///     instance to make sure all queued and in-flight produce requests are
-        ///     completed before terminating. 
-        ///    
-        ///     A related configuration parameter is message.timeout.ms which determines
-        ///     the maximum length of time librdkafka attempts to deliver a message 
-        ///     before giving up and so also affects the maximum time a call to Flush 
-        ///     may block.
-        /// 
-        ///     Where this Producer instance shares a Handle with one or more other
-        ///     producer instances, the Flush method will wait on messages produced by
-        ///     the other producer instances as well.
-        /// </remarks>
-        /// <exception cref="System.OperationCanceledException">
-        ///     Thrown if the operation is cancelled.
-        /// </exception>
-        public void Flush(CancellationToken cancellationToken = default(CancellationToken))
-            => core.Flush(cancellationToken);
-
-        /// <summary>
-        ///     <see cref="IClient.AddBrokers(string)" />
-        /// </summary>
-        public int AddBrokers(string brokers)
-            => core.AddBrokers(brokers);
-
-        /// <summary>
-        ///     Releases all resources used by this <see cref="Producer{TKey,TValue}" />.
-        /// </summary>
-        public void Dispose()
-            => core.Dispose();
 
         private class TypedTaskDeliveryHandlerShim<K, V> : TaskCompletionSource<DeliveryResult<K, V>>, IDeliveryHandler
         {
