--- src/Confluent.Kafka/ProducerCore.cs
@@ -1,475 +1,0 @@
-// Copyright 2016-2018 Confluent Inc., 2015-2016 Andreas Heider
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Derived from: rdkafka-dotnet, licensed under the 2-clause BSD License.
-//
-// Refer to LICENSE for more information.
-
-using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using Confluent.Kafka.Impl;
-using Confluent.Kafka.Internal;
-
-
-namespace Confluent.Kafka
-{
-    /// <summary>
-    ///     A high-level Apache Kafka producer, excluding any public methods
-    ///     for producing messages.
-    /// </summary>
-    internal class ProducerCore : IClient
-    {
-        internal class Config
-        {
-            internal IEnumerable<KeyValuePair<string, string>> config;
-            internal Action<Error> errorHandler;
-            internal Action<LogMessage> logHandler;
-            internal Action<string> statisticsHandler;
-        }
-        
-        private int cancellationDelayMaxMs;
-        private bool disposeHasBeenCalled = false;
-        private object disposeHasBeenCalledLockObj = new object();
-
-        private bool manualPoll = false;
-        internal bool enableDeliveryReports = true;
-        internal bool enableDeliveryReportKey = true;
-        internal bool enableDeliveryReportValue = true;
-        internal bool enableDeliveryReportTimestamp = true;
-        internal bool enableDeliveryReportHeaders = true;
-        internal bool enableDeliveryReportPersistedStatus = true;
-
-        private SafeKafkaHandle ownedKafkaHandle;
-        private Handle borrowedHandle;
-
-        internal SafeKafkaHandle KafkaHandle
-            => ownedKafkaHandle != null 
-                ? ownedKafkaHandle
-                : borrowedHandle.LibrdkafkaHandle;
-
-        private Task callbackTask;
-        private CancellationTokenSource callbackCts;
-
-        private Task StartPollTask(CancellationToken ct)
-            => Task.Factory.StartNew(() =>
-                {
-                    try
-                    {
-                        while (true)
-                        {
-                            ct.ThrowIfCancellationRequested();
-                            ownedKafkaHandle.Poll((IntPtr)cancellationDelayMaxMs);
-                        }
-                    }
-                    catch (OperationCanceledException) {}
-                }, ct, TaskCreationOptions.LongRunning, TaskScheduler.Default);
-
-
-        private Action<Error> errorHandler;
-        private Librdkafka.ErrorDelegate errorCallbackDelegate;
-        private void ErrorCallback(IntPtr rk, ErrorCode err, string reason, IntPtr opaque)
-        {
-            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
-            if (ownedKafkaHandle.IsClosed) { return; }
-            errorHandler?.Invoke(KafkaHandle.CreatePossiblyFatalError(err, reason));
-        }
-
-
-        private Action<string> statisticsHandler;
-        private Librdkafka.StatsDelegate statisticsCallbackDelegate;
-        private int StatisticsCallback(IntPtr rk, IntPtr json, UIntPtr json_len, IntPtr opaque)
-        {
-            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
-            if (ownedKafkaHandle.IsClosed) { return 0; }
-            statisticsHandler?.Invoke(Util.Marshal.PtrToStringUTF8(json));
-            return 0; // instruct librdkafka to immediately free the json ptr.
-        }
-
-
-        private Action<LogMessage> logHandler;
-        private object loggerLockObj = new object();
-        private Librdkafka.LogDelegate logCallbackDelegate;
-        private void LogCallback(IntPtr rk, SyslogLevel level, string fac, string buf)
-        {
-            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
-            // Note: kafkaHandle can be null if the callback is during construction (in that case, we want the delegate to run).
-            if (ownedKafkaHandle != null && ownedKafkaHandle.IsClosed) { return; }
-            logHandler?.Invoke(new LogMessage(Util.Marshal.PtrToStringUTF8(Librdkafka.name(rk)), level, fac, buf));
-        }
-
-        private Librdkafka.DeliveryReportDelegate DeliveryReportCallback;
-
-        private void DeliveryReportCallbackImpl(IntPtr rk, IntPtr rkmessage, IntPtr opaque)
-        {
-            // Ensure registered handlers are never called as a side-effect of Dispose/Finalize (prevents deadlocks in common scenarios).
-            if (ownedKafkaHandle.IsClosed) { return; }
-
-            var msg = Util.Marshal.PtrToStructureUnsafe<rd_kafka_message>(rkmessage);
-
-            // the msg._private property has dual purpose. Here, it is an opaque pointer set
-            // by Topic.Produce to be an IDeliveryHandler. When Consuming, it's for internal
-            // use (hence the name).
-            if (msg._private == IntPtr.Zero)
-            {
-                // Note: this can occur if the ProduceAsync overload that accepts a DeliveryHandler
-                // was used and the delivery handler was set to null.
-                return;
-            }
-
-            var gch = GCHandle.FromIntPtr(msg._private);
-            var deliveryHandler = (IDeliveryHandler) gch.Target;
-            gch.Free();
-
-            Headers headers = null;
-            if (this.enableDeliveryReportHeaders) 
-            {
-                headers = new Headers();
-                Librdkafka.message_headers(rkmessage, out IntPtr hdrsPtr);
-                if (hdrsPtr != IntPtr.Zero)
-                {
-                    for (var i=0; ; ++i)
-                    {
-                        var err = Librdkafka.header_get_all(hdrsPtr, (IntPtr)i, out IntPtr namep, out IntPtr valuep, out IntPtr sizep);
-                        if (err != ErrorCode.NoError)
-                        {
-                            break;
-                        }
-                        var headerName = Util.Marshal.PtrToStringUTF8(namep);
-                        byte[] headerValue = null;
-                        if (valuep != IntPtr.Zero)
-                        {
-                            headerValue = new byte[(int)sizep];
-                            Marshal.Copy(valuep, headerValue, 0, (int)sizep);
-                        }
-                        headers.Add(headerName, headerValue);
-                    }
-                }
-            }
-
-            IntPtr timestampType = (IntPtr)TimestampType.NotAvailable;
-            long timestamp = 0;
-            if (enableDeliveryReportTimestamp)
-            {
-                timestamp = Librdkafka.message_timestamp(rkmessage, out timestampType);
-            }
-
-            PersistenceStatus messageStatus = PersistenceStatus.PossiblyPersisted;
-            if (enableDeliveryReportPersistedStatus)
-            {
-                messageStatus = Librdkafka.message_status(rkmessage);
-            }
-
-            deliveryHandler.HandleDeliveryReport(
-                new DeliveryReport<Null, Null>
-                {
-                    // Topic is not set here in order to avoid the marshalling cost.
-                    // Instead, the delivery handler is expected to cache the topic string.
-                    Partition = msg.partition, 
-                    Offset = msg.offset, 
-                    Error = KafkaHandle.CreatePossiblyFatalError(msg.err, null),
-                    Message = new Message<Null, Null> { Timestamp = new Timestamp(timestamp, (TimestampType)timestampType), Headers = headers }
-                }
-            );
-        }
-
-        public void Produce(
-            string topic,
-            byte[] val, int valOffset, int valLength,
-            byte[] key, int keyOffset, int keyLength,
-            Timestamp timestamp,
-            Partition partition, 
-            IEnumerable<Header> headers,
-            IDeliveryHandler deliveryHandler)
-        {
-            if (timestamp.Type != TimestampType.CreateTime)
-            {
-                if (timestamp != Timestamp.Default)
-                {
-                    throw new ArgumentException("Timestamp must be either Timestamp.Default, or Timestamp.CreateTime.");
-                }
-            }
-
-            if (this.enableDeliveryReports && deliveryHandler != null)
-            {
-                // Passes the TaskCompletionSource to the delivery report callback via the msg_opaque pointer
-
-                // Note: There is a level of indirection between the GCHandle and
-                // physical memory address. GCHandle.ToIntPtr doesn't get the
-                // physical address, it gets an id that refers to the object via
-                // a handle-table.
-                var gch = GCHandle.Alloc(deliveryHandler);
-                var ptr = GCHandle.ToIntPtr(gch);
-
-                var err = KafkaHandle.Produce(
-                    topic,
-                    val, valOffset, valLength,
-                    key, keyOffset, keyLength,
-                    partition.Value,
-                    timestamp.UnixTimestampMs,
-                    headers,
-                    ptr);
-
-                if (err != ErrorCode.NoError)
-                {
-                    gch.Free();
-                    throw new KafkaException(KafkaHandle.CreatePossiblyFatalError(err, null));
-                }
-            }
-            else
-            {
-                var err = KafkaHandle.Produce(
-                    topic,
-                    val, valOffset, valLength,
-                    key, keyOffset, keyLength,
-                    partition.Value,
-                    timestamp.UnixTimestampMs,
-                    headers,
-                    IntPtr.Zero);
-
-                if (err != ErrorCode.NoError)
-                {
-                    throw new KafkaException(KafkaHandle.CreatePossiblyFatalError(err, null));
-                }
-            }
-        }
-
-        public ProducerCore(Handle handle)
-        {
-            if (!handle.Owner.GetType().Name.Contains("Producer")) // much simpler than checking actual types.
-            {
-                throw new Exception("A Producer instance may only be constructed using the handle of another Producer instance.");
-            }
-
-            this.borrowedHandle = handle;
-        }
-
-        public ProducerCore(ProducerCore.Config baseConfig)
-        {
-            // TODO: Make Tasks auto complete when EnableDeliveryReportsPropertyName is set to false.
-            // TODO: Hijack the "delivery.report.only.error" configuration parameter and add functionality to enforce that Tasks 
-            //       that never complete are never created when this is set to true.
-
-            this.statisticsHandler = baseConfig.statisticsHandler;
-            this.logHandler = baseConfig.logHandler;
-            this.errorHandler = baseConfig.errorHandler;
-
-            var config = Confluent.Kafka.Config.ExtractCancellationDelayMaxMs(baseConfig.config, out this.cancellationDelayMaxMs);
-
-            this.DeliveryReportCallback = DeliveryReportCallbackImpl;
-
-            Librdkafka.Initialize(null);
-
-            var modifiedConfig = config
-                .Where(prop => 
-                    prop.Key != ConfigPropertyNames.Producer.EnableBackgroundPoll &&
-                    prop.Key != ConfigPropertyNames.Producer.EnableDeliveryReports &&
-                    prop.Key != ConfigPropertyNames.Producer.DeliveryReportFields);
-
-            if (modifiedConfig.Where(obj => obj.Key == "delivery.report.only.error").Count() > 0)
-            {
-                // A managed object is kept alive over the duration of the produce request. If there is no
-                // delivery report generated, there will be a memory leak. We could possibly support this 
-                // property by keeping track of delivery reports in managed code, but this seems like 
-                // more trouble than it's worth.
-                throw new ArgumentException("The 'delivery.report.only.error' property is not supported by this client");
-            }
-
-            var enableBackgroundPollObj = config.FirstOrDefault(prop => prop.Key == ConfigPropertyNames.Producer.EnableBackgroundPoll).Value;
-            if (enableBackgroundPollObj != null)
-            {
-                this.manualPoll = !bool.Parse(enableBackgroundPollObj.ToString());
-            }
-
-            var enableDeliveryReportsObj = config.FirstOrDefault(prop => prop.Key == ConfigPropertyNames.Producer.EnableDeliveryReports).Value;
-            if (enableDeliveryReportsObj != null)
-            {
-                this.enableDeliveryReports = bool.Parse(enableDeliveryReportsObj.ToString());
-            }
-
-            var deliveryReportEnabledFieldsObj = config.FirstOrDefault(prop => prop.Key == ConfigPropertyNames.Producer.DeliveryReportFields).Value;
-            if (deliveryReportEnabledFieldsObj != null)
-            {
-                var fields = deliveryReportEnabledFieldsObj.ToString().Replace(" ", "");
-                if (fields != "all")
-                {
-                    this.enableDeliveryReportKey = false;
-                    this.enableDeliveryReportValue = false;
-                    this.enableDeliveryReportHeaders = false;
-                    this.enableDeliveryReportTimestamp = false;
-                    this.enableDeliveryReportPersistedStatus = false;
-                    if (fields != "none")
-                    {
-                        var parts = fields.Split(',');
-                        foreach (var part in parts)
-                        {
-                            switch (part)
-                            {
-                                case "key": this.enableDeliveryReportKey = true; break;
-                                case "value": this.enableDeliveryReportValue = true; break;
-                                case "timestamp": this.enableDeliveryReportTimestamp = true; break;
-                                case "headers": this.enableDeliveryReportHeaders = true; break;
-                                case "status": this.enableDeliveryReportPersistedStatus = true; break;
-                                default: throw new ArgumentException(
-                                    $"Unknown delivery report field name '{part}' in config value '{ConfigPropertyNames.Producer.DeliveryReportFields}'.");
-                            }
-                        }
-                    }
-                }
-            }
-
-            var configHandle = SafeConfigHandle.Create();
-
-            modifiedConfig.ToList().ForEach((kvp) => {
-                if (kvp.Value == null) throw new ArgumentException($"'{kvp.Key}' configuration parameter must not be null.");
-                configHandle.Set(kvp.Key, kvp.Value.ToString());
-            });
-
-
-            IntPtr configPtr = configHandle.DangerousGetHandle();
-
-            if (enableDeliveryReports)
-            {
-                Librdkafka.conf_set_dr_msg_cb(configPtr, DeliveryReportCallback);
-            }
-
-            // Explicitly keep references to delegates so they are not reclaimed by the GC.
-            errorCallbackDelegate = ErrorCallback;
-            logCallbackDelegate = LogCallback;
-            statisticsCallbackDelegate = StatisticsCallback;
-
-            // TODO: provide some mechanism whereby calls to the error and log callbacks are cached until
-            //       such time as event handlers have had a chance to be registered.
-            Librdkafka.conf_set_error_cb(configPtr, errorCallbackDelegate);
-            Librdkafka.conf_set_log_cb(configPtr, logCallbackDelegate);
-            Librdkafka.conf_set_stats_cb(configPtr, statisticsCallbackDelegate);
-
-            this.ownedKafkaHandle = SafeKafkaHandle.Create(RdKafkaType.Producer, configPtr, this);
-            configHandle.SetHandleAsInvalid(); // config object is no longer useable.
-
-            if (!manualPoll)
-            {
-                callbackCts = new CancellationTokenSource();
-                callbackTask = StartPollTask(callbackCts.Token);
-            }
-        }
-
-        public int Poll(TimeSpan timeout)
-        {
-            if (!manualPoll)
-            {
-                throw new InvalidOperationException("Poll method called, but manual polling is not enabled.");
-            }
-
-            return this.KafkaHandle.Poll((IntPtr)timeout.TotalMillisecondsAsInt());
-        }
-
-        public int Flush(TimeSpan timeout)
-            => KafkaHandle.Flush(timeout.TotalMillisecondsAsInt());
-
-
-        public void Flush(CancellationToken cancellationToken)
-        {
-            while (true)
-            {
-                int result = KafkaHandle.Flush(100);
-                if (result == 0)
-                {
-                    return;
-                }
-                if (cancellationToken.IsCancellationRequested)
-                {
-                    // TODO: include flush number in exception.
-                    throw new OperationCanceledException();
-                }
-            }
-        }
-        
-        public void Dispose()
-        {
-            Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            // Calling Dispose a second or subsequent time should be a no-op.
-            lock (disposeHasBeenCalledLockObj)
-            { 
-                if (disposeHasBeenCalled) { return; }
-                disposeHasBeenCalled = true;
-            }
-
-            // do nothing if we borrowed a handle.
-            if (ownedKafkaHandle == null) { return; }
-
-            if (disposing)
-            {
-                if (!this.manualPoll)
-                {
-                    callbackCts.Cancel();
-                    try
-                    {
-                        // Note: It's necessary to wait on callbackTask before disposing kafkaHandle
-                        // since the poll loop makes use of this.
-                        callbackTask.Wait();
-                    }
-                    catch (AggregateException e)
-                    {
-                        if (e.InnerException.GetType() != typeof(TaskCanceledException))
-                        {
-                            throw e.InnerException;
-                        }
-                    }
-                    finally
-                    {
-                        callbackCts.Dispose();
-                    }
-                }
-
-                // calls to rd_kafka_destroy may result in callbacks
-                // as a side-effect. however the callbacks this class
-                // registers with librdkafka ensure that any registered
-                // events are not called if kafkaHandle has been closed.
-                // this avoids deadlocks in common scenarios.
-                ownedKafkaHandle.Dispose();
-            }
-        }
-
-        public string Name
-            => KafkaHandle.Name;
-
-        public int AddBrokers(string brokers)
-            => KafkaHandle.AddBrokers(brokers);
-
-        public Handle Handle 
-        {
-            get
-            {
-                if (this.ownedKafkaHandle != null)
-                {
-                    return new Handle { Owner = this, LibrdkafkaHandle = ownedKafkaHandle };
-                }
-                
-                return borrowedHandle;
-            }
-        }
-
-    }
-}
